Move the process metrics collector code into its own file
